requirements.txt
flask
flask-cors
pandas
scikit-learn
scikit-surprise
numpy

